+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#192] Add missing clientSet-dependency to ManagerSet
+  - This fixes a bug when removing component service-accounts
 
 ## [v2.0.1] - 2024-08-08
 ### Fixed